The entire _Installers folder cross-compiles net48/netcoreapp3.1, resulting in multiple additional files being included in the .NET Core compilation. A few stub types allow for compilation and will be replaced with real implementations in separate tickets.
