--- Application/EdFi.Ods.AdminApp.Web.Core/Controllers/StubControllers.cs
@@ -1,19 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Licensed to the Ed-Fi Alliance under one or more agreements.
-// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
-// See the LICENSE and NOTICES files in the project root for more information.
-
-namespace EdFi.Ods.AdminApp.Web.Controllers
-{
-    public class OdsInstanceSettingsController
-    {
-        public object Applications() => throw new System.NotImplementedException();
-        public object Descriptors() => throw new System.NotImplementedException();
-        public object Logging() => throw new System.NotImplementedException();
-        public object EducationOrganizations() => throw new System.NotImplementedException();
-        public object Setup() => throw new System.NotImplementedException();
-        public object BulkLoad() => throw new System.NotImplementedException();
-        public object LearningStandards() => throw new System.NotImplementedException();
-        public object SelectDistrict(int i) => throw new System.NotImplementedException();
-    }
-}
