+++ Application/EdFi.Ods.AdminApp.Web.Core/EdFi.Ods.AdminApp.Web.Core.csproj
@@ -17,9 +17,9 @@
          these files will be moved from the legacy project to this project, at
          which point these explicit inclusions will become redundant. -->
 
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/_Installers/AzureInstaller.cs" LinkBase="_Installers" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/_Installers/CommonConfigurationInstaller.cs" LinkBase="_Installers" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/_Installers/OnPremInstaller.cs" LinkBase="_Installers" /> -->
+    <Compile Include="../EdFi.Ods.AdminApp.Web/_Installers/AzureInstaller.cs" LinkBase="_Installers" />
+    <Compile Include="../EdFi.Ods.AdminApp.Web/_Installers/CommonConfigurationInstaller.cs" LinkBase="_Installers" />
+    <Compile Include="../EdFi.Ods.AdminApp.Web/_Installers/OnPremInstaller.cs" LinkBase="_Installers" />
 
     <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/ActionFilters/AuthorizationFilter.cs" LinkBase="ActionFilters" /> -->
     <Compile Include="../EdFi.Ods.AdminApp.Web/ActionFilters/BypassInstanceContextFilter.cs" LinkBase="ActionFilters" />
@@ -109,10 +109,10 @@
     <Compile Include="../EdFi.Ods.AdminApp.Web/Helpers/StringExtensions.cs" LinkBase="Helpers" />
     <Compile Include="../EdFi.Ods.AdminApp.Web/Helpers/VendorExtensions.cs" LinkBase="Helpers" />
 
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Hubs/BulkUploadHub.cs" LinkBase="Hubs" /> -->
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Hubs/BulkUploadHub.cs" LinkBase="Hubs" />
     <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Hubs/EdfiOdsHub.cs" LinkBase="Hubs" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Hubs/ProductionLearningStandardsHub.cs" LinkBase="Hubs" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Hubs/ProductionSetupHub.cs" LinkBase="Hubs" /> -->
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Hubs/ProductionLearningStandardsHub.cs" LinkBase="Hubs" />
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Hubs/ProductionSetupHub.cs" LinkBase="Hubs" />
 
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/AcceptAttribute.cs" LinkBase="Infrastructure" />
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/AutoMapper/AdminWebMappingProfile.cs" LinkBase="Infrastructure/AutoMapper" />
@@ -123,7 +123,7 @@
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/CloudOdsAdminAppSettings.cs" LinkBase="Infrastructure" />
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/CloudOdsAdminAppSettingsForAzure.cs" LinkBase="Infrastructure" />
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/CloudOdsApiConnectionInformationProvider.cs" LinkBase="Infrastructure" />
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/CloudOdsAzureActiveDirectoryClientInfo.cs" LinkBase="Infrastructure" /> -->
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/CloudOdsAzureActiveDirectoryClientInfo.cs" LinkBase="Infrastructure" />
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/CloudOdsUpdateService.cs" LinkBase="Infrastructure" />
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/EducationOrganizationTypes.cs" LinkBase="Infrastructure" />
     <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/HangFire/HangFireInstance.cs" LinkBase="Infrastructure/HangFire" /> -->
@@ -135,11 +135,11 @@
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/IO/InterchangeFileHelpers.cs" LinkBase="Infrastructure/IO" />
     <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/IO/InterchangeFileType.cs" LinkBase="Infrastructure/IO" />
     <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/IO/LocalFileSystemFileUploadHandler.cs" LinkBase="Infrastructure/IO" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/BulkUploadJob.cs" LinkBase="Infrastructure/Jobs" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/IWorkflowJob.cs" LinkBase="Infrastructure/Jobs" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/LearningStandardsJob.cs" LinkBase="Infrastructure/Jobs" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/ProductionSetupJob.cs" LinkBase="Infrastructure/Jobs" /> -->
-    <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/WorkflowJob.cs" LinkBase="Infrastructure/Jobs" /> -->
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/BulkUploadJob.cs" LinkBase="Infrastructure/Jobs" />
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/IWorkflowJob.cs" LinkBase="Infrastructure/Jobs" />
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/LearningStandardsJob.cs" LinkBase="Infrastructure/Jobs" />
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/ProductionSetupJob.cs" LinkBase="Infrastructure/Jobs" />
+    <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/Jobs/WorkflowJob.cs" LinkBase="Infrastructure/Jobs" />
     <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/ModelBinding/CompositeModelBinder.cs" LinkBase="Infrastructure/ModelBinding" /> -->
     <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/ModelBinding/EnumerationModelBinder.cs" LinkBase="Infrastructure/ModelBinding" /> -->
     <!-- <Compile Include="../EdFi.Ods.AdminApp.Web/Infrastructure/ModelBinding/EnumerationPropertyBinder.cs" LinkBase="Infrastructure/ModelBinding" /> -->
